Change the language to support the REPL.
This includes support for eval and null.

`eval` is parsed and compiled as an expression statement with one
argument, which can of course be an expression. This results in the
expression being evaluated to the Z value, followed by an `EVAL` opcode.

`null` is a sigil value, added now as a return value from the `read`
stdlib function to indicate EOF.

Neither `read` nor `eval` are implemented yet.